requirements.txt
arrow==0.13.0
atomicwrites==1.3.0
attrs==18.2.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
entrypoints==0.3
flake8==3.7.5
idna==2.8
lxml==4.3.0
mccabe==0.6.1
more-itertools==5.0.0
pluggy==0.8.1
py==1.7.0
pycodestyle==2.5.0
pyflakes==2.1.0
pytest==4.2.0
python-dateutil==2.8.0
requests==2.21.0
six==1.12.0
soupsieve==1.7.3
tinydb==3.12.2
urllib3==1.24.1
